# sandboxd-orch/docs/orchestrator_swagger.yaml
basePath: /
definitions:
...
  handlers.DeleteSandboxResponse:
    properties:
      deleted:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.SandboxObjectResponse:
    properties:
      sandbox:
        $ref: '#/definitions/types.Sandbox'
    type: object
  handlers.SandboxObjectsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/types.Sandbox'
        type: array
    type: object
...
  types.CreateSandboxObjectRequest:
    properties:
      id:
        type: string
      spec:
        $ref: '#/definitions/types.SandboxSpec'
    type: object
...
  types.Sandbox:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      spec:
        $ref: '#/definitions/types.SandboxSpec'
      status:
        $ref: '#/definitions/types.SandboxStatus'
      updated_at:
        type: string
    type: object
  types.SandboxContainerSpec:
    properties:
      args:
        items:
          type: string
        type: array
      env:
        items:
          type: string
        type: array
      image:
        type: string
      name:
        type: string
      resource:
        $ref: '#/definitions/types.SandboxResource'
      work_dir:
        type: string
    type: object
  types.SandboxPhase:
    enum:
    - Pending
    - Scheduled
    - Running
    - Failed
    - Deleting
    type: string
    x-enum-varnames:
    - SandboxPhasePending
    - SandboxPhaseScheduled
    - SandboxPhaseRunning
    - SandboxPhaseFailed
    - SandboxPhaseDeleting
  types.SandboxPortAssign:
    properties:
      container_port:
        type: integer
      host_port:
        type: integer
      protocol:
        type: string
    type: object
  types.SandboxPortSpec:
    properties:
      container_port:
        type: integer
      host_port:
        type: integer
      protocol:
        type: string
    type: object
  types.SandboxResource:
    properties:
      cpu:
        type: string
      memory:
        type: string
    type: object
  types.SandboxSpec:
    properties:
      containers:
        items:
          $ref: '#/definitions/types.SandboxContainerSpec'
        type: array
      egress:
        type: boolean
      ports:
        items:
          $ref: '#/definitions/types.SandboxPortSpec'
        type: array
      ttl_seconds:
        type: integer
    type: object
  types.SandboxStatus:
    properties:
      assigned_ports:
        items:
          $ref: '#/definitions/types.SandboxPortAssign'
        type: array
      expire_at:
        type: string
      last_error:
        type: string
      node_name:
        type: string
      phase:
        $ref: '#/definitions/types.SandboxPhase'
      sandboxd_id:
        type: string
    type: object
info:
  contact: {}
  description: Control-plane API for node registration, health probing, sandbox scheduling
    objects, and node proxy operations.
  title: Orchestrator API
  version: "1.0"
paths:
...
  /api/v1/sandboxes:
    get:
      description: Lists orchestrator sandbox objects and their scheduling status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SandboxObjectsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List control-plane sandboxes
      tags:
      - orchestrator-sandbox
    post:
      consumes:
      - application/json
      description: Creates a scheduler-managed sandbox object in orchestrator.
      parameters:
      - description: Control-plane sandbox request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.CreateSandboxObjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.SandboxObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create control-plane sandbox object
      tags:
      - orchestrator-sandbox
  /api/v1/sandboxes/{id}:
    delete:
      description: Deletes orchestrator sandbox object and requests runtime deletion
        on assigned node.
      parameters:
      - description: Sandbox ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DeleteSandboxResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete control-plane sandbox
      tags:
      - orchestrator-sandbox
    get:
      description: Returns one orchestrator sandbox object.
      parameters:
      - description: Sandbox ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SandboxObjectResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get control-plane sandbox
      tags:
      - orchestrator-sandbox
...
schemes:
- http
swagger: "2.0"